Refactor auth/auth.go in package crypto
